rename layers in workspaces
