+++ services/web/server/src/simcore_service_webserver/projects/projects_api.py
@@ -120,7 +120,7 @@ from ..users.preferences_api import (
 )
 from ..wallets import api as wallets_api
 from ..wallets.errors import WalletNotEnoughCreditsError
-from ..workspaces import _workspaces_db as workspaces_db
+from ..workspaces import _workspaces_repository as workspaces_db
 from . import _crud_api_delete, _nodes_api, _projects_db
 from ._access_rights_api import (
     check_user_project_permission,
