+++ services/web/server/src/simcore_service_webserver/workspaces/api.py
@@ -1,5 +1,5 @@
 # mypy: disable-error-code=truthy-function
-from ._workspaces_api import check_user_workspace_access, get_workspace
+from ._workspaces_service import check_user_workspace_access, get_workspace
 
 assert get_workspace  # nosec
 assert check_user_workspace_access  # nosec
